+++ jim_hats_mobile/lib/presentation/views/ranking/ranking_page.dart
@@ -46,7 +46,7 @@ class RankingView extends StatelessWidget {
             return SafeArea(
                 child: Padding(
                     padding: EdgeInsets.symmetric(
-                        horizontal: AppSpacings.horizontalPadding.toDouble()),
+                        horizontal: AppSpacings.horizontalPadding),
                     child: ListView(
                       children: [
                         Padding(
