+++ jim_hats_mobile/lib/presentation/views/welcome/welcome_page.dart
@@ -16,7 +16,7 @@ class WelcomePage extends StatelessWidget {
             alignment: Alignment.bottomCenter,
             child: Padding(
               padding: EdgeInsets.symmetric(
-                  horizontal: AppSpacings.horizontalPadding.toDouble()),
+                  horizontal: AppSpacings.horizontalPadding),
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.stretch,
                 mainAxisSize: MainAxisSize.min,
